# Small speedup by simplifying BitWriter::send_bits