@folkertdev 2024-12-08
Hi, 

For benchmarking it is crucial that you actually race the two implementations against each other: just comparing wall clock times is unreliable, especially when the times are so close together (and in fact their ranges overlap).

When I run this commit agains its base commit, I actually observe a regression

```
Benchmark 1 (65 runs): ./compress-baseline 1 rs silesia-small.tar
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          77.2ms ±  844us    76.2ms … 82.6ms          4 ( 6%)        0%
  peak_rss           26.7MB ± 64.3KB    26.6MB … 26.7MB          0 ( 0%)        0%
  cpu_cycles          283M  ± 2.02M      279M  …  292M           4 ( 6%)        0%
  instructions        633M  ±  277       633M  …  633M           0 ( 0%)        0%
  cache_references   20.0M  ±  162K     19.7M  … 20.7M           2 ( 3%)        0%
  cache_misses        466K  ± 67.0K      374K  …  735K           2 ( 3%)        0%
  branch_misses      3.06M  ± 7.05K     3.05M  … 3.09M           1 ( 2%)        0%
Benchmark 2 (64 runs): ./target/release/examples/blogpost-compress 1 rs silesia-small.tar
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          79.1ms ±  422us    78.5ms … 80.3ms          2 ( 3%)        💩+  2.5% ±  0.3%
  peak_rss           26.7MB ± 68.6KB    26.5MB … 26.7MB          0 ( 0%)          -  0.0% ±  0.1%
  cpu_cycles          292M  ± 1.31M      290M  …  297M           2 ( 3%)        💩+  3.3% ±  0.2%
  instructions        659M  ±  300       659M  …  659M           1 ( 2%)        💩+  4.1% ±  0.0%
  cache_references   19.9M  ±  190K     19.6M  … 20.5M           1 ( 2%)          -  0.8% ±  0.3%
  cache_misses        455K  ± 79.0K      353K  …  723K           1 ( 2%)          -  2.3% ±  5.4%
  branch_misses      3.11M  ± 2.09K     3.11M  … 3.12M           0 ( 0%)        💩+  1.6% ±  0.1%
```

(I'm using https://github.com/andrewrk/poop here)

The branch that you remove is not actually conditional in practice: it is very well-predicted. While it doesn't look like it, this code is actually fairly optimized and it's hard to do better (I've tried!). 

So, I guess, see what you get when you run the two benchmarks alongside each other (hyperfine accepts multiple things to benchmark, you just need to copy the binary from before your change somewhere so you can use it after applying your change).