@brian-pane 2024-12-08
Thanks, I'll adopt the same benchmark method. In my initial test run using the same benchmark driver, I see an improvement in branch mispredictions, although it's offset by an increase in instructions executed:
```
poop "./compress-baseline 1 rs silesia-small.tar" "./target/release/examples/compress 1 rs silesia-small.tar" 
Benchmark 1 (55 runs): ./compress-baseline 1 rs silesia-small.tar
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          91.6ms ± 1.51ms    90.3ms …  101ms          4 ( 7%)        0%
  peak_rss           26.7MB ± 67.0KB    26.5MB … 26.7MB          0 ( 0%)        0%
  cpu_cycles          341M  ±  706K      340M  …  343M           0 ( 0%)        0%
  instructions        748M  ±  234       748M  …  748M           0 ( 0%)        0%
  cache_references    399K  ± 2.02K      396K  …  406K           3 ( 5%)        0%
  cache_misses        294K  ± 8.94K      266K  …  313K           6 (11%)        0%
  branch_misses      3.28M  ± 4.31K     3.27M  … 3.29M           0 ( 0%)        0%
Benchmark 2 (56 runs): ./target/release/examples/compress 1 rs silesia-small.tar
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          90.5ms ±  541us    88.6ms … 92.0ms          6 (11%)          -  1.2% ±  0.5%
  peak_rss           26.7MB ± 64.8KB    26.5MB … 26.7MB          0 ( 0%)          +  0.0% ±  0.1%
  cpu_cycles          339M  ±  578K      338M  …  341M           0 ( 0%)          -  0.6% ±  0.1%
  instructions        771M  ±  257       771M  …  771M           0 ( 0%)        💩+  3.1% ±  0.0%
  cache_references    398K  ± 2.55K      396K  …  412K           7 (13%)          -  0.1% ±  0.2%
  cache_misses        293K  ± 9.10K      266K  …  304K           9 (16%)          -  0.5% ±  1.2%
  branch_misses      3.20M  ± 10.6K     3.18M  … 3.22M           0 ( 0%)        ⚡-  2.4% ±  0.1%
```

I suspect the difference in branch mispredictions compared to your results might be due to running on different CPU models. But I haven't yet figured out why my patch causes an increase in instructions. I'll look at it more later.