@brian-pane 2024-12-12
I tried a different approach in [76b6c14](https://github.com/trifectatechfoundation/zlib-rs/pull/258/commits/76b6c14b634c677a89dbe6e97897f2bbc63f8168): keeping the original logic, but inlining the fast path. It seems to work on my test system.