@brian-pane 2024-12-12
I just tried with `#[cold]` and the results look basically unchanged on my system.